feat: implemented option to change your room booking to anorther room in the same hotel
